# docs/openapi-v2.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "剧院死神后端",
    "version": "2.2",
    "description": "数字UID登录、可改昵称、直接注册与管理员账号管理契约。API版本2、玩法规则2.0。"
  },
  "servers": [
    {
      "url": "/api/v2"
    }
  ],
  "security": [
    {
      "accountCookie": []
    }
  ],
  "paths": {
...
    "/auth/login": {
      "parameters": [],
      "post": {
        "summary": "登录（不自动接管房间）",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "uid": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "uid",
                  "password"
                ],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMe"
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "parameters": [],
      "post": {
        "summary": "注销当前会话；控制会话注销时离开房间",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "required": [],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "loggedOut": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "loggedOut"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "失败。按error.code处理；未知结果不能视为操作已失败。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/avatars/{assetId}": {
      "parameters": [
        {
          "name": "assetId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "summary": "登录后读取规范化头像",
        "responses": {
          "200": {
            "description": "静态WebP",
            "content": {
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "304": {
            "description": "缓存仍有效"
          },
          "default": {
            "description": "无权限或资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "livekitSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "LiveKit服务端签名；SDK验证原始body。"
      },
      "accountCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "td_account_v2",
        "description": "HttpOnly会话，浏览器自动携带。隔离开发3003使用td_account_contract_21，按服务配置区分。"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "AuthMe": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "uid": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "profileVersion": {
            "type": "integer"
          },
          "expiresAt": {
            "type": "integer"
          }
        },
        "required": [
          "userId",
          "uid",
          "nickname",
          "avatarUrl",
          "profileVersion",
          "expiresAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
}